requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
docopt==0.6.2
gidgethub==4.2.0
idna==2.10
irc3==1.1.6
limnoria==2020.10.10
multidict==5.0.2
openttd-helpers==1.0.1
pycparser==2.20
PyJWT==1.7.1
PyYAML==5.3.1
sentry-sdk==0.19.4
six==1.15.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
venusian==3.0.0
yarl==1.6.3
